requirements.txt
discord.py==2.7.3
fastapi==0.111.1
uvicorn==0.23.2
jinja2==3.1.3
python-multipart==0.0.6
